requirements.txt
djangorestframework==3.14.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
gunicorn==22.0.0